requirements.txt
fastapi[all]
psycopg2[all]
pyjwt[all]
passlib[all]
python-dotenv[all]
pyjwt